vsock: flow control fix
Fixed a logical error in the vsock flow control, that would cause credit
update packets to not be sent at the right time.

+++ src/devices/src/virtio/vsock/csm/txbuf.rs
@@ -22,7 +22,7 @@ pub struct TxBuf {
 
 impl TxBuf {
     /// Total buffer size, in bytes.
-    const SIZE: usize = defs::CONN_TX_BUF_SIZE;
+    const SIZE: usize = defs::CONN_TX_BUF_SIZE as usize;
 
     /// Ring-buffer constructor.
     pub fn new() -> Self {
